requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
lightgbm==4.5.0
xgboost==2.1.1
pydantic==2.8.2
python-dateutil==2.9.0.post0
joblib==1.4.2
mlflow==2.16.2
evidently==0.5.5
pyyaml==6.0.2

prefect==3.0.12
google-cloud-storage==2.18.2

requests==2.31.0
holidays==0.34
cachetools==5.3.0
pytz==2025.7

evidently==0.5.5
slack-sdk==3.22.0
optuna==3.3.0
flask==2.3.4
react-scripts==5.0.1
